+++ toolbox/fdc3-workbench/src/fixtures/logMessages.ts
@@ -19,6 +19,10 @@ export const getLogMessage = (name: logMessagesName, type: logMessagesType, valu
       success: `Retrieved current channel [${value}]`,
       error: `Failed to retrieve current channel`,
     },
+    userChannelChanged: {
+      info: `User channel changed to [${value}]`,
+      error: `Failed to listen for user channel changes`,
+    },
     joinUserChannel: {
       success: `Joined the [${value}] channel`,
       error: `Failed to join the [${value}] channel`,
@@ -111,6 +115,9 @@ export const getLogMessage = (name: logMessagesName, type: logMessagesType, valu
     pcOnDisconnect: {
       success: `${value}`,
       error: `${value}`,
+    },
+    privateChannelEventListener: {
+      error: `Failed to listen for events on private channel [${value}]`,
     },
   };